+++ backend/for_society/for_society.hpp
@@ -7,6 +7,6 @@ template <uint64_t from_not, uint64_t to_not>
 uint64_t FromNotToNot(uint64_t) noexcept;
 
 // extract digit from integer by index
-uint8_t GetDigit(uint8_t number, uint8_t index) noexcept;
+uint32_t GetDigit(uint32_t number, uint32_t index) noexcept;
 
 }  // namespace FS
